requirements.txt
boto3==1.16.10
fastapi==0.61.1
dnspython==2.0.0
pymongo==3.11.0
pycryptodome==3.9.8
nanoid==2.0.0
python-multipart==0.0.5
uvicorn==0.12.2
